[mod/#38] id-> Long으로 수정

+++ app/src/main/java/com/haphap/app/presentation/search/SearchScreen.kt
@@ -61,7 +61,7 @@ private fun SearchScreen(
     onRelatedItemClick: (Int) -> Unit,
     onFilterClick: (String) -> Unit,
     onResultCardClick: (Int) -> Unit,
-    onDeleteClick: (Int) -> Unit,
+    onDeleteClick: (Long) -> Unit,
     onTrendCardClick: (Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
